// packages/discord/src/operations/deprecatedListPins.ts
export const DeprecatedListPinsInput =
  /*@__PURE__*/ /*#__PURE__*/ Schema.Struct({
    channel_id: Schema.String.pipe(T.PathParam()),
  }).pipe(T.Http({ method: "GET", path: "/channels/{channel_id}/pins" }));
export type DeprecatedListPinsInput = typeof DeprecatedListPinsInput.Type;
// ...
export type DeprecatedListPinsOutput = typeof DeprecatedListPinsOutput.Type;

// The operation
export const deprecatedListPins = /*@__PURE__*/ /*#__PURE__*/ API.make(() => ({
  inputSchema: DeprecatedListPinsInput,
  outputSchema: DeprecatedListPinsOutput,
  errors: [BadRequest, Forbidden, NotFound] as const,
}));